+++ src/shared/components/TemplateItem/index.tsx
