+++ src/shared/components/TemplateItem/style.ts
